server/Server_DMPG/requirements.txt
simpy~=4.1.1
pandas~=2.2.1
numpy~=1.26.4
matplotlib~=3.8.3
seaborn~=0.13.2
graphviz~=2.50.0
setuptools~=68.2.2
sqlalchemy
psycopg2
request
